+++ rx-runtime-reaktive/settings.gradle.kts
@@ -1,0 +1,2 @@
+rootProject.name = "rx-runtime-reaktive"
+enableFeaturePreview("GRADLE_METADATA")
