+++ rx-runtime-reaktive/src/androidMain/AndroidManifest.xml
