+++ rx-runtime-reaktive/src/androidTest/resources/.keep
