+++ rx-runtime-reaktive/src/commonMain/kotlin/com/suparnatural/runtime/rx/reaktive/Reaktive.kt
@@ -1,4 +1,4 @@
-package com.suparnatural.runtime.rx
+package com.suparnatural.runtime.rx.reaktive
 
 import com.badoo.reaktive.maybe.asObservable
 import com.badoo.reaktive.observable.*
@@ -9,17 +9,17 @@ import com.suparnatural.core.rx.PublishSubject
 import com.suparnatural.core.rx.PublishSubjectFactory
 
 
-class BadooObservable<T>(private val observable: com.badoo.reaktive.observable.Observable<T>) : Observable<T> {
+class ReaktiveObservable<T>(private val observable: com.badoo.reaktive.observable.Observable<T>) : Observable<T> {
     override fun <V> map(project: (value: T) -> V): Observable<V> {
-        return BadooObservable(observable.map(project))
+        return ReaktiveObservable(observable.map(project))
     }
 
     override fun filter(predicate: (value: T) -> Boolean): Observable<T> {
-        return BadooObservable(observable.filter(predicate))
+        return ReaktiveObservable(observable.filter(predicate))
     }
 
     override fun reduce(accumulator: (acc: T, value: T) -> T, seed: T?): Observable<T> {
-        return BadooObservable(observable.reduce(accumulator).asObservable())
+        return ReaktiveObservable(observable.reduce(accumulator).asObservable())
     }
 
     override fun subscribe(subscriber: (value: T) -> Unit) {
@@ -27,9 +27,9 @@ class BadooObservable<T>(private val observable: com.badoo.reaktive.observable.O
     }
 }
 
-class BadooPublishSubject<T>(private val subject: com.badoo.reaktive.subject.publish.PublishSubject<T>) : PublishSubject<T> {
+class ReaktivePublishSubject<T>(private val subject: com.badoo.reaktive.subject.publish.PublishSubject<T>) : PublishSubject<T> {
     override fun asObservable(): Observable<T> {
-        return BadooObservable(subject)
+        return ReaktiveObservable(subject)
     }
 
     override fun onNext(value: T) {
@@ -45,14 +45,14 @@ class BadooPublishSubject<T>(private val subject: com.badoo.reaktive.subject.pub
     }
 }
 
-class BadooObservableFactory : ObservableFactory {
+class ReaktiveObservableFactory : ObservableFactory {
     override fun <T> of(vararg values: T): Observable<T> {
-        return BadooObservable(observableOf(*values))
+        return ReaktiveObservable(observableOf(*values))
     }
 }
 
-class BadooPublishSubjectFactory : PublishSubjectFactory {
+class ReaktivePublishSubjectFactory : PublishSubjectFactory {
     override fun <T> create(): PublishSubject<T> {
-        return BadooPublishSubject(publishSubject())
+        return ReaktivePublishSubject(publishSubject())
     }
 }
