+++ rx-runtime-reaktive/src/iosTest/resources/Info.plist
