--- rx-runtime/settings.gradle.kts
@@ -1,2 +1,0 @@
-rootProject.name = "rx-runtime"
-enableFeaturePreview("GRADLE_METADATA")
